corregido Adapter y renombradas las actividades
