+++ app/src/main/AndroidManifest.xml
@@ -23,13 +23,13 @@
         <activity
             android:name=".LoginActivity"
             android:label="@string/title_activity_login" />
-        <activity android:name=".Work" />
-        <activity android:name=".Proyects" />
-        <activity android:name=".Habits" />
-        <activity android:name=".toDoList" />
-        <activity android:name=".Goals" />
-        <activity android:name=".Character" />
-        <activity android:name=".Stats"></activity>
+        <activity android:name=".WorkActivity" />
+        <activity android:name=".ProyectsActivity" />
+        <activity android:name=".HabitSActivity" />
+        <activity android:name=".toDoListActivity" />
+        <activity android:name=".GoalsActivity" />
+        <activity android:name=".CharacterActivity" />
+        <activity android:name=".StatsActivity"></activity>
     </application>
 
 </manifest>
