+++ app/src/main/java/com/example/milagros/improof/CharacterActivity.java
@@ -3,11 +3,11 @@ package com.example.milagros.improof;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class Stats extends AppCompatActivity {
+public class CharacterActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_stats);
+        setContentView(R.layout.activity_character);
     }
 }
