--- app/src/main/java/com/example/milagros/improof/Habits.java
@@ -1,13 +1,0 @@
-package com.example.milagros.improof;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class Habits extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_habits);
-    }
-}
