+++ app/src/main/java/com/example/milagros/improof/HabitsActivity.java
@@ -1,0 +1,13 @@
+package com.example.milagros.improof;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class HabitsActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_habits);
+    }
+}
