--- app/src/main/java/com/example/milagros/improof/Proyects.java
@@ -1,14 +1,0 @@
-package com.example.milagros.improof;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class Proyects extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_proyects);
-    }
-    
-}
