+++ app/src/main/java/com/example/milagros/improof/StatsActivity.java
@@ -3,11 +3,11 @@ package com.example.milagros.improof;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class Goals extends AppCompatActivity {
+public class StatsActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_goals);
+        setContentView(R.layout.activity_stats);
     }
 }
