+++ app/src/main/res/layout/activity_character.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.milagros.improof.Character">
+    tools:context="com.example.milagros.improof.CharacterActivity">
     <ImageButton
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
