+++ app/src/main/res/layout/activity_habits.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.milagros.improof.Habits">
+    tools:context="com.example.milagros.improof.HabitsActivity">
     <ListView
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
