+++ app/src/main/res/layout/activity_proyects.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.milagros.improof.Proyects">
+    tools:context="com.example.milagros.improof.ProyectsActivity">
     <ListView
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
