+++ app/src/main/res/layout/activity_stats.xml
@@ -8,6 +8,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.milagros.improof.Stats">
+    tools:context="com.example.milagros.improof.StatsActivity">
 
 </RelativeLayout>
