+++ app/src/main/res/layout/activity_to_do_list.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:orientation="vertical"
-    tools:context="com.example.milagros.improof.toDoList">
+    tools:context="com.example.milagros.improof.toDoListActivity">
     <ListView
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
